Fix so that search box retains focus on each keystroke.
Each key stroke would trigger a new API request for blocks. While this
was loading, the loading spinner would be shown. Then when the response
was received, the blocks list would update but the search box would no longer have focus.

This change avoids the loading spinner being shown for a microsecond and so the search box
does not lose focus.