Link product name and image to product page.
Previously there was no way to get to a product page. This restores the same functionality
as you would expect from a carousel or grid that was created through stencil.

The GraphQL queries have been updated to include a request for the site URL which
can then be passed through the components.

The product card has then been updated to show the overlay buttons on hover of just
the image container which then has the linked image and the buttons within.

I've reduced the buttons to auto width rather than 100% to give a little more space
to click the image when hovering. They're still not quite as small as on cornerstone
but better I think.

The way I've optionally defined a href on links could definitely be improved but it works.